+++ src/memories/Memory.java
@@ -1,4 +1,4 @@
-package models;
+package memories;
 
 import java.io.File;
