+++ src/processes/UnixProcess.java
@@ -1,25 +1,26 @@
-package models;
+package processes;
 
-public class WindowProcess extends ProcessInfo {
-  public final static String[] TITLES = { "User", "Status", "Name", "Pid", "%Cpu", "Mem" };
+public class UnixProcess extends ProcessInfo {
+  private final static String KILL_PROCESS_COMMAND = "kill -9 ";
+  public final static String[] TITLES = { "User", "Status", "Name", "Pid", "%Cpu", "%Mem" };
   private static final int POSITION_USER = 0;
-  private static final int POSITION_STATUS = 1;
-  private static final int POSITION_PID = 2;
+  private static final int POSITION_PID = 1;
+  private static final int POSITION_CPU = 2;
   private static final int POSITION_MEM = 3;
-  private static final int POSITION_CPU = 4;
-  private static final int POSITION_NAME = 5;
-  private final String KILL_PROCESS_COMMAND = "taskkill ";
+  private static final int POSITION_STATUS = 7;
+  private static final int POSITION_NAME = 10;
 
   @Override
   protected String getKillProcessCommand() {
     return KILL_PROCESS_COMMAND;
   }
 
-  public WindowProcess() {
+  public UnixProcess() {
     super();
   }
 
-  public WindowProcess(String string) {
+  public UnixProcess(String string) {
+    super();
     String[] values = string.split(" ");
     String user = values[POSITION_USER];
     String status = values[POSITION_STATUS];
@@ -30,7 +31,8 @@ public class WindowProcess extends ProcessInfo {
     setAttributes(user, status, name, id, cpu, mem);
   }
 
-  public WindowProcess(String user, String status, String name, long id, float cpu, float mem) {
+  public UnixProcess(String user, String status, String name, long id, float cpu, float mem) {
     super(user, status, name, id, cpu, mem);
   }
+
 }
