+++ src/processes/WindowProcesses.java
@@ -1,31 +1,36 @@
-package models;
+package processes;
 
 import java.io.BufferedReader;
 import java.io.InputStreamReader;
 import java.util.ArrayList;
 
-public class UnixProcesses extends ProcessesManagement {
-  private static String LIST_PROCESSES_COMMAND = "ps aux --sort -rss";
+public class WindowProcesses extends ProcessesManagement {
+  private final String GET_LIST_PROCESSES_COMMAND = "tasklist ";
 
-  public UnixProcesses(ArrayList<ProcessInfo> processes) {
+  public WindowProcesses(ArrayList<ProcessInfo> processes) {
     super(processes);
   }
 
-  public UnixProcesses() {
+  public WindowProcesses() {
     super();
+  }
+
+  public String getListProcessesCommand() {
+    return GET_LIST_PROCESSES_COMMAND;
   }
 
   @Override
   public ArrayList<ProcessInfo> loadProcesses() {
     try {
-      Process p = Runtime.getRuntime().exec(LIST_PROCESSES_COMMAND);
+      Process p = Runtime.getRuntime().exec(getListProcessesCommand());
       BufferedReader input = new BufferedReader(new InputStreamReader(p.getInputStream()));
       String line;
       processes.clear();
-//      remove the first line which is the titles of the processes
+      // Remove 2 first lines which is the titles of the processes in Window OS
+      input.readLine();
       input.readLine();
       while ((line = input.readLine()) != null) {
-        processes.add(new UnixProcess(breeze(line)));
+        processes.add(new WindowProcess(breeze(line)));
       }
       input.close();
       return processes;
@@ -37,7 +42,6 @@ public class UnixProcesses extends ProcessesManagement {
 
   @Override
   public String[] getTitles() {
-    return UnixProcess.TITLES;
+    return WindowProcess.TITLES;
   }
-
 }
